fix: don't install aznfs on ARM64 (#6481)

+++ parts/linux/cloud-init/artifacts/ubuntu/cse_install_ubuntu.sh
@@ -51,7 +51,9 @@ installDeps() {
     fi
 
     if [ "${OSVERSION}" = "22.04" ] || [ "${OSVERSION}" = "24.04" ]; then
-        pkg_list+=("aznfs=0.3.15")
+        if [ "$(isARM64)" -eq 0 ]; then
+            pkg_list+=("aznfs=0.3.15")
+        fi
     fi
 
     for apt_package in ${pkg_list[*]}; do
@@ -62,9 +64,11 @@ installDeps() {
     done
 
     if [ "${OSVERSION}" = "22.04" ] || [ "${OSVERSION}" = "24.04" ]; then
-    	# disable aznfswatchdog since aznfs install and enable aznfswatchdog and aznfswatchdogv4 services at the same time while we only need aznfswatchdogv4
-        systemctl disable aznfswatchdog
-        systemctl stop aznfswatchdog
+        if [ "$(isARM64)" -eq 0 ]; then
+            # disable aznfswatchdog since aznfs install and enable aznfswatchdog and aznfswatchdogv4 services at the same time while we only need aznfswatchdogv4
+            systemctl disable aznfswatchdog
+            systemctl stop aznfswatchdog
+        fi
     fi
 }
